requirements.txt
# System prerequisites (install once via apt):
#   sudo apt install -y swig liblgpio-dev python3-dev python3-venv

fastapi
uvicorn[standard]
opencv-python-headless
ultralytics
gpiozero
lgpio
python-dotenv
